LayoutTests/fast/events/touch/ios/show-modal-alert-during-touch-start-expected.txt

ALERT: This is a modal alert.
Tap to show an alert
This test verifies that presenting a modal alert while handling a synchronous touch event doesn't cause the application process to hang. To verify manually, tap the red area above, and check that the JavaScript alert is presented with no significant delay.
